Windows linker fixes for release builds.
Now Debug and Release versions both build fine on Windows.
I'm still not satisfied that getting mpirxx to work with xank and stl is
so much of a bother, but I'll worry about this later. I think we can do
fine without gmp C++ version and just work with C and things are (as always) fine.